backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-restful
marshmallow-sqlalchemy
flask-marshmallow
psycopg2-binary
python-dotenv
# celery
# redis
passlib
